create context folder and move user context file in
